requirements.txt
flask>=3.0.0
flask_wtf>=1.2.1
pillow>=10.1.0
psutil>=5.9.6
pylibmc>=1.6.3
pytest>=7.4.3
rms-pdslogger>=0.0.5
rms-pdstable>=0.0.5
